Add a `showExpand` prop to SearchField component
Setting `showExpand` on `SearchField` as `true` causes an icon to appear right-justified in the search input. When the user clicks on that icon, a popup appears with a text field that the user can type in and hit "Search" to submit the form (submission behavior is passed via a new `onSubmit` prop).

This also adds a new svg icon to the assets repo since the icon we need does not exist in the material-ui icons pack.