CHECK-1116 Confirmation on navigation inside app
When the user is editing an annotation form, `router.setRouteLeaveHook` is called which presents a custom message and confirmation dialog to the user asking if they want to continue and lose their unsaved work. This moves the `isEditing` state up in the component tree from `Tasks` to `MediaPageLayout` so that it can be shared between `Tasks` and `NextOrPreviousLinks`.