Fixing rerender bug in CHECK-1302
The number field was losing focus every time the user pressed a key to enter a value. It was actually rerendering with each keystroke because we were rendering the filter component from a function called from within a render cycle itself. By moving the DateRangeSelector components outside of the DateRangeFilter component, passing in the needed values, we avoid the rerender. More here https://stackoverflow.com/a/60510217

Also split `DateRangeSelector` into two functions, `DateRangeSelectorStartEnd` and `DateRangeSelectorRelative`, and moved the conditional to the render of the main component itself.

Made some changes to strings as requested, too.